+++ runtime/integration-tests/src/generic/mod.rs
@@ -1,0 +1,11 @@
+pub mod env;
+pub mod envs {
+	pub mod fudge_env;
+	pub mod runtime_env;
+}
+pub mod runtimes;
+
+// Test cases
+mod cases {
+	mod example;
+}
