+++ runtime/integration-tests/src/generic/runtimes.rs
@@ -1,7 +1,7 @@
 use frame_support::inherent::InherentData;
 use sp_runtime::{traits::Block, ApplyExtrinsicResult};
 
-use super::env::{Config, RuntimeKind};
+use crate::generic::env::{Config, RuntimeKind};
 
 impl Config for development_runtime::Runtime {
 	type Block = development_runtime::Block;
@@ -25,4 +25,30 @@ impl Config for development_runtime::Runtime {
 	fn finalize_block() -> <Self::Block as Block>::Header {
 		development_runtime::Executive::finalize_block()
 	}
+}
+
+#[macro_export]
+macro_rules! test_with_all_runtimes {
+	($setup:ident, $name:ident) => {
+		mod $name {
+			use super::*;
+
+			#[test]
+			fn development() {
+				$setup::<development_runtime::Runtime>()
+					.execute_with($name::<development_runtime::Runtime>);
+			}
+
+			#[test]
+			fn altair() {
+				$setup::<altair_runtime::Runtime>().execute_with($name::<altair_runtime::Runtime>);
+			}
+
+			#[test]
+			fn centrifuge() {
+				$setup::<centrifuge_runtime::Runtime>()
+					.execute_with($name::<centrifuge_runtime::Runtime>);
+			}
+		}
+	};
 }
