--- runtime/integration-tests/src/option_b/mod.rs
@@ -1,5 +1,0 @@
-pub mod env;
-pub mod runtimes;
-pub mod test_env;
-
-mod foo;
